Added "ResendableContent"

+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendAnimation.kt
@@ -17,7 +17,7 @@ fun SendAnimation(
     thumb: InputFile?,
     caption: String? = null,
     parseMode: ParseMode? = null,
-    duration: Int? = null,
+    duration: Long? = null,
     width: Int? = null,
     height: Int? = null,
     disableNotification: Boolean = false,
@@ -71,7 +71,7 @@ data class SendAnimationData internal constructor(
     override val parseMode: ParseMode? = null,
     @SerialName(durationField)
     @Optional
-    override val duration: Int? = null,
+    override val duration: Long? = null,
     @SerialName(widthField)
     @Optional
     override val width: Int? = null,
